Fix WeightedEdge definition
Although conceptually correct, having WeightedEdge and TwoWeightedEdge class definitions for this
problem set is likely to lead to problems, as the automated grader presumes that what I previously
called TwoWeightedEdge is actually called WeightedEdge, and that what I previously called WeightedEdge
is not defined for this problem set.

Accordingly, delete WeightedEdge, and rename TwoWeightedEdge to WeightedEdge.