Replace weights with descriptive names
While conceptually it makes sense to implement a more-general WeightedEdge class, for the purposes
of this program, it is expected and understood that the first weight is the totalDistance, while
the second weight is the outsideDistance.

Accordingly, delete reference to and use of wname{1,2}, replace weight1 with totalDistance,
and replace weight2 with outsideDistance.  Also, update the getWeight{1,2} methods to reflect this
new naming scheme.