Add implementation for WeightedDigraph
In the Digraph class, edges are stored as a dictionary of sources pointing to a list of destinations.
However, this does not include the weight information that we would want to store along with that source-
destination pair.  Additionally, that weight information is immutable for that source-destination pair.

That suggests that the ideal data structure would be a tuple, containing the (destination, total, outside)
information, which would replace the destination entry in the list of possible destinations for that
particular source node.

Additionally, the __str__ method needs to be extended for WeightedDigraph to produce, along with the
source-destination pairs, an output of the associated weights for that connection.  This is done by
referencing the individual elements of the tuple in order.