Replace 'outside' with 'outdoor' and implement WeightedDigraph.childrenOf
The problem set presumes that the word outdoor is used rather than the word outside.

Additionally, as the weighted edges are stored as a tuple, we must go deeper to access just the
desired destination for a given source node.  Hence, append [0] to the node call in the childrenOf()
method for the WeightedDigraph class.