Use dictionary to store nodes
There is a compounding problem of sequence in PS5.  That is to say, the I/O iterates over the flat
text file line by line.  In doing this, it learns about new nodes (if any), and also about a new edge.
To accomplish this, however, it has to recall nodes it's already seen.  It's presumed that each line is
a new edge (that is, an edge it hasn't seen before).

Further compounding this problem, node objects created from the names seen in the flat file have to be kept
track of throughout the iteration through the file, or else these nodes can't be referenced at the end of the
iteration, when the edge is added to the WeightedDigraph.

The solution to this is to use a dictionary, whereby knowing the name of the node allows you to reference
the associated node object.