Reduce iteration, and need for temporary lists
Previously, the approach to this function was to iterate through the file line by line, then iterate
through each stored entry in nodeDict to add to the WeightedDigraph, then iterate through each stored
entry in listOfEdges to add to the WeightedDigraph.

While logically sound, this approach requires ever-larger storage for the listOfEdges as the flat file
size grows.  Unfortunately, nodeDict still is required to grow along with the flat file size, but a
reduction is useful even if it's not a complete reduction.

The solution is to use the existing has-node-at-column-0-been-added-to-nodeDict to also add that node
to the WeightedDigraph.  This same extension was also made for the node-at-column-1, so that it is known
that both nodes exist in the WeightedDigraph by the end of these checks.  As this is the case, we have
no further reservations about adding the edge that references both of those nodes.  Accordingly, the edge
represented by that line of the flat file is now added at the end of the line-iteration.

Maybe nodeDict could be done away with entirely, thus causing only WeightedDigraph to grow along with the
size of the flat file, though I doubt it, and at any rate it's expedient to be able to call '45' when you
mean 'the node which was created when the iteration happened upon the node value of 45'.  As such, it's
probably good enough at present.