Enable finding a node with just node name.
The node storage for Digraph is a set, but unfortunately just stores node objects.
in ps5.py, bruteForceSearch will be called using strings for the start and end node,
rather than the node objects themselves.  This is a problem, as there is no persistent
mapping for node names to node objects (one is created in ps5.load_map, but this dictionary
isn't stored anywhere after that function call ends).

The most logical place for such a mapping would be the WeightedDigraph itself, and as
discussed previously, a dictionary makes good sense as the data structure to accomplish
this.

As such, methods defined in WeightedDigraph's superclass such as __init__,
addNode, and hasNode, must be rewritten to accomodate this change in data structure.
Additionally, since callers may have the nodeName, but not the node object, it
makes sense to define new methods getNode, which returns the node object associated
with that node name, and hasNodeName, which provides a quicker route for callers to
check whether the node exists in the WeightedDigraph than hasNode(getNode(nodeName)).