Revert to earlier node data structure, but maintain lookup
There are times within this program when a function having only the name of
a node needs to access that node object, thus it made sense to have a
dictionary data structure which was able to provide node objects when given
the node name.

Unfortunately, the program requires that the WeightedDigraph.node data structure
is a set, not a dictionary.

Thus, the solution is to keep WeightedDigraph.node as a set(), but to construct
a separate dictionary data structure WeightedDigraph.nodeTable, which is able
to take in strings corresponding to node names and provide node object, as
desired.