Fix WeightedDigraph.childrenOf
The data structure of WeightedDigraph.edges changed to accomodate storing weights.
childrenOf() had not been updated to reflect this change.

Specifically, Digraph.edges is a dictionary of Nodes:list pairs, where the Nodes
are the keys and the lists are the values.  These lists in turn store solely Nodes.
Put differently, for each Source Node (key), there is a list of Destination Nodes
as the matching value.

However, in order to incorporate values associated with changing state between start
and end nodes, the Nodes:list pair has been further extended to a data structure
as follows: Start: [[destNode1, (w1, w2)], [destNode2, (w3, w4)], [destNode3, (w5, w5)]].
Accordingly, if a list of destination Nodes is desired, it is now necessary to
iterate through the uberList, then for each entry, to pull the first index value
and store that in a temporary list (children).

Accordingly, implement iteration and temporary value storage, before a list of
children nodes is returned.