Revise load_map and implement bruteForceSearch
As the data structure for WeightedDigraph now contains a dictionary of
nodeName:Node pairs, such a dictionary is now unnecessary within
load_map itself.  Thus, this dictionary and its use have been removed.

Relatedly, as WeightedDigraph has implemented Node return based on a
given nodeName string, it is no longer necessary to for load_map to
store Node objects so that they can be referenced later in graph
construction.  So, load_map now needs only ask WeightedDigraph whether
it's got a node that goes by a given nodeName string.  If not, it's then
able to create a Node and add it to WeightedDigraph in a single step
(WeightedDigraph handles stashing the name:Obj pair).

bruteForceSearch required some revision from the lecture code,
specifically because it's being passed strings to start rather than node
objects (and thus, recursive calls to bFS also have to pass strings).
This isn't a big deal though, since it just means asking WeightedDigraph
for the node objects, then using those Node objects as before, but
passing bruteForceSearch the corresponding string when needed.