Add methods for edges
Edges are a useful data structure, but nowhere is the Edge object stored.
Accordingly, implement an edgeTable to solve this problem, which makes use of
(sourceNode, destinationNode) tuples as keys, and Edge objects as values.
Thus, the WeightedDigraph is able to return the edge object that corresponds
to a particular (source, destination) ordered pair.  This is useful since the
Edge data structure is able to return information about the costs of changing
from source state to destination state.

Additionally, implement methods to return the total and outdoor distances
from a WeightedDigraph object given a path (a list of nodes), making use of the
above methods and data structures.