Fix ValueError logic
bruteForceSearch() is a recursive algorithm, which returns either a list
or None.  When bruteForceSearch is first called on a graph, it will make
several recursive calls to itself before it unwinds and returns any
values.

On the first run, if it's gotten to the end of the path it's searching, but hasn't yet
reached the specified end (goal) node, it will return None.  This is
because it's not returning path, and there are no nodes in
digraph.childrenOf(startNode) [because it's at the end], so it just
returns shortest, which is still None.

Eventually, one of the calls of bruteForceSearch() will be to the end
node, so startNode == endNode and that call of bruteForceSearch() will
return path.

That returned value, though, still has to be evaluated by the remaining
logic in the preceding bruteForceSearch() call.  Specifically, it will
pass the newPath != None test (because it's a valid path), but here we
might also want to check to see if that path satisfies both of the given
weight constraints before saving it to shortest.  Doing it this way
allows us to only save paths to shortest which are both valid and which
satisfy the constraints.

Knowing that the only non-None values for shortest that can be returned
are paths that reach from the start node to the end node, and which
satisfy both given constraints, we know that, if we have unwound all the
way back to the first call of bruteForceSearch, and shortest is still
None, either there is no path between start and end, or any valid paths
do not satisfy the given constraints, or both.  We don't care about
valid paths that don't satisfy the constraints.

A neat trick to aid in this logic is that, len(path) is 1 when and only
when you first call bruteForceSearch.  Accordingly, before returning
shortest, we should check to see if we're in the base case.  If we're in
the base case where shortest = None, we don't meet the program
specifications. But, if we're not in the base case, returning a None
shortest is totally fine, and aids in our logic.

Accordingly, if it is the case that we're in the base case and shortest
is None, we know that no path satisfies the given constraints, and we
can safely raise a ValueError.