Fix logic to only overwrite shortest when shorter
Previously, the code would overwrite shortest whenever a newPath that was found
satisfied the constraints given to the problem.  This had the unfortunate
consequence that a newPath found later in the recursion would overwrite a
path found earlier in the recursion, without checking whether that newPath was
shorter (total distance) than the earlier, existing path.

Accordingly, the logic prose is as follows:  If bFS() returned a path that meets
both given constraints, that path is our shortest path if we don't have a
shortest path already.  If we do have a shortest path, that new path overwrites
our shortest path if its total distance is less than the total distance of the
existing shortest path.