Copy bruteForce to directedDFS
It appears that, rather than implementing bruteForceSearch()
as the problem wanted it, I instead implemented to
bruteForceSearch() what the problem described as
directedDFS().  Accordingly, directedDFS() is done, and
has been copied and tested.

However, the task of implementing bruteForceSearch()
remains.  I think the first part of this task will be to
understand what is meant by bruteForceSearch.  In the
context of weighted edge pathfinding, this probably means
an implementation of breadthFirstSearch, since we don't
actually know anything from the mere depth of the search,
but rather we must consult the appropriate weights.