Lobotomize bruteForceSearch
Previously, bruteForceSearch had logic checks that prevented it from
investigating a further node if it knows that the length of the path
it's currently on already exceeds the best path found.

However, that removes the "brute force" aspect from this algorithm, and
instead makes it directed.  Accordingly, in order to make the algorithm
behave as-desired, these checks need to be removed.

The variable shortest will still store the shortest path found thus far,
and other no-path ValueError checks will still function as before, but
removing the above checks ensures that all possible paths are explored.