Added bruteForcePruneSearch function
This function uses the highly-inefficient bruteForceSearch function to develop further efficiency
in the recursion.  Essentially, once the function notices that it's either about to enter a loop
state, or enter a state where there are dead-end nodes (that is, nodes that don't reach the
goal node), it stops the recursion.